requirements.txt
crewai
crewai-tools
requests
pydantic
litellm
python-dotenv
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-multipart
Pillow
